Refactoring: split files, combine interfaces into TaskEx
